p4i integration Integrate 260302 from 4.3 to main: Fixes:    16 -> 8 bit color conversion Task:     159018 RevBy:    none, TrustMe
[git-p4: depot-paths = "//depot/qt/main/": change = 260303]

+++ src/gui/painting/qdrawhelper_p.h
@@ -254,7 +254,7 @@ inline quint8 qt_colorConvert(quint32 color)
 template <>
 inline quint8 qt_colorConvert(quint16 color)
 {
-    return qt_colorConvert<quint32, quint8>(qt_conv16ToRgb(color));
+    return qt_colorConvert<quint8, quint32>(qt_conv16ToRgb(color));
 }
 #endif // QT_QWS_DEPTH_8
